app/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
passlib==1.7.4
